basic storybook component working
