+++ package/src/forwardedProps.ts
@@ -1,0 +1,1 @@
+export type { ButtonProps as MyBLDRButtonProps } from "@mui/material/Button";
